requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
torch==2.5.1
torchvision==0.20.1
Pillow==11.0.0
requests==2.26.0
numpy==2.0.2
streamlit==1.12.0
streamlit-drawable-canvas==0.9.3
altair==4.2.0
python-multipart

